Finish prediction parsing
