+++ mosqlient/errors.py
@@ -13,7 +13,9 @@ class FieldTypeError(ValidationErrorBase):
 
     def __init__(self, field: str, _type: Union[Iterable, Any]):
         self.field = field
-        super().__init__(f"Incorrect type for field: '{field}'. Expected type(s): '{_type}'")
+        super().__init__(
+            f"Incorrect type for field: '{field}'. Expected type(s): '{_type}'"
+        )
 
     def __str__(self):
         return f"{self.field}: {self.args[0]}"
@@ -27,3 +29,8 @@ class ClientError(Exception):
 class ModelPostError(Exception):
     def __init__(self, message=None):
         super().__init__(message or "registry.Model POST request error")
+
+
+class PredictionPostError(Exception):
+    def __init__(self, message=None):
+        super().__init__(message or "registry.Prediction POST request error")
